if (patternMatcher.find()) geändert zu while (patternMatcher.find()) is one of pattern um Varianten ergänzt lemmatitzer ergänzt float-Gewichtung eingeführt diverse Strukturanpassungen Wortfilter erweitert

+++ src/main/java/TaxonomySearcher.java
@@ -1,7 +1,7 @@
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
@@ -32,7 +32,7 @@ public class TaxonomySearcher extends Configured implements Tool {
         job.setReducerClass(WeightReducer.class);
 
         job.setOutputKeyClass(Text.class);
-        job.setOutputValueClass(IntWritable.class);
+        job.setOutputValueClass(FloatWritable.class);
 
         FileInputFormat.addInputPath(job, new Path(args[0]));
         FileOutputFormat.setOutputPath(job, new Path(args[1]));
