+++ src/main/java/WeightReducer.java
@@ -1,15 +1,15 @@
-import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
-public class WeightReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
-    private IntWritable result = new IntWritable();
+public class WeightReducer extends Reducer<Text, FloatWritable, Text, FloatWritable> {
+    private FloatWritable result = new FloatWritable();
 
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
-        for (IntWritable val : values) {
+    public void reduce(Text key, Iterable<FloatWritable> values, Context context) throws IOException, InterruptedException {
+        float sum = 0;
+        for (FloatWritable val : values) {
             sum += val.get();
         }
         result.set(sum);
